feat(dashboard): unified list items for deployments
fixed the way the latest scheduled or pending deployment is tracked
