+++ dashboard/src/components/common/DeploymentListItem/index.ts
@@ -1,0 +1,2 @@
+export * from './DeploymentListItem';
+export { default } from './DeploymentListItem';
