+++ dashboard/src/components/overview/OverviewDeployments/OverviewDeployments.tsx
@@ -1,32 +1,21 @@
-import { AppDeploymentDuration } from '@/components/applications/AppDeployments';
 import { EditRepositorySettings } from '@/components/applications/github/EditRepositorySettings';
 import useGitHubModal from '@/components/applications/github/useGitHubModal';
+import DeploymentListItem from '@/components/common/DeploymentListItem';
 import { useDialog } from '@/components/common/DialogProvider';
 import NavLink from '@/components/common/NavLink';
 import GithubIcon from '@/components/icons/GithubIcon';
-import Tooltip from '@/components/ui/v2/Tooltip';
 import { useCurrentWorkspaceAndApplication } from '@/hooks/useCurrentWorkspaceAndApplication';
-import { Avatar } from '@/ui/Avatar';
-import Status, { StatusEnum } from '@/ui/Status';
-import type { DeploymentStatus } from '@/ui/StatusCircle';
-import { StatusCircle } from '@/ui/StatusCircle';
 import ActivityIndicator from '@/ui/v2/ActivityIndicator';
 import Button from '@/ui/v2/Button';
-import ArrowCounterclockwiseIcon from '@/ui/v2/icons/ArrowCounterclockwiseIcon';
 import RocketIcon from '@/ui/v2/icons/RocketIcon';
-import type { ListItemRootProps } from '@/ui/v2/ListItem';
-import { ListItem } from '@/ui/v2/ListItem';
+import List from '@/ui/v2/List';
 import Text from '@/ui/v2/Text';
 import { getLastLiveDeployment } from '@/utils/helpers';
-import { toastStyleProps } from '@/utils/settings/settingsConstants';
-import type { DeploymentRowFragment } from '@/utils/__generated__/graphql';
 import {
   useGetDeploymentsSubSubscription,
-  useInsertDeploymentMutation,
+  useScheduledOrPendingDeploymentsSubSubscription,
 } from '@/utils/__generated__/graphql';
 import { ChevronRightIcon } from '@heroicons/react/solid';
-import { formatDistanceToNowStrict, parseISO } from 'date-fns';
-import { toast } from 'react-hot-toast';
 import { twMerge } from 'tailwind-merge';
 
 function OverviewDeploymentsTopBar() {
@@ -61,156 +50,6 @@ function OverviewDeploymentsTopBar() {
   );
 }
 
-interface OverviewDeploymentProps extends ListItemRootProps {
-  /**
-   * Deployment metadata to display.
-   */
-  deployment: DeploymentRowFragment;
-  /**
-   * Determines to show a status badge showing the live status of a deployment reflecting the latest state of the application.
-   */
-  isDeploymentLive: boolean;
-  /**
-   * Determines whether or not the redeploy button should be shown for the
-   * deployment.
-   */
-  showRedeploy?: boolean;
-  /**
-   * Determines whether or not the redeploy button is disabled.
-   */
-  disableRedeploy?: boolean;
-}
-
-function OverviewDeployment({
-  deployment,
-  isDeploymentLive,
-  className,
-  showRedeploy,
-  disableRedeploy,
-}: OverviewDeploymentProps) {
-  const { currentWorkspace, currentApplication } =
-    useCurrentWorkspaceAndApplication();
-
-  const showTime =
-    !['SCHEDULED', 'PENDING'].includes(deployment.deploymentStatus) &&
-    deployment.deploymentStartedAt;
-
-  const relativeDateOfDeployment = showTime
-    ? formatDistanceToNowStrict(parseISO(deployment.deploymentStartedAt), {
-        addSuffix: true,
-      })
-    : '';
-
-  const [insertDeployment, { loading }] = useInsertDeploymentMutation();
-
-  const { commitMessage } = deployment;
-
-  return (
-    <ListItem.Root className={className}>
-      <ListItem.Button
-        className="grid grid-flow-col items-center justify-between gap-2 px-2 py-2"
-        component={NavLink}
-        href={`/${currentWorkspace.slug}/${currentApplication.slug}/deployments/${deployment.id}`}
-      >
-        <div className="flex cursor-pointer flex-row items-center justify-center space-x-2 self-center">
-          <Avatar
-            name={deployment.commitUserName}
-            avatarUrl={deployment.commitUserAvatarUrl}
-            className="h-8 w-8 shrink-0"
-          />
-
-          <div className="grid grid-flow-row truncate text-sm+ font-medium">
-            <Text className="inline cursor-pointer truncate font-medium leading-snug text-greyscaleDark">
-              {commitMessage?.trim() || (
-                <span className="truncate pr-1 font-normal italic">
-                  No commit message
-                </span>
-              )}
-            </Text>
-
-            <Text className="text-sm font-normal leading-[1.375rem] text-greyscaleGrey">
-              {relativeDateOfDeployment}
-            </Text>
-          </div>
-        </div>
-
-        <div className="grid grid-flow-col gap-2 items-center">
-          {showRedeploy && (
-            <Tooltip
-              title="An active deployment cannot be re-triggered"
-              hasDisabledChildren={disableRedeploy || loading}
-              disableHoverListener={!disableRedeploy}
-            >
-              <Button
-                disabled={disableRedeploy || loading}
-                size="small"
-                color="secondary"
-                variant="outlined"
-                onClick={async (event) => {
-                  event.stopPropagation();
-                  event.preventDefault();
-
-                  const insertDeploymentPromise = insertDeployment({
-                    variables: {
-                      object: {
-                        appId: currentApplication?.id,
-                        commitMessage: deployment.commitMessage,
-                        commitSHA: deployment.commitSHA,
-                        commitUserAvatarUrl: deployment.commitUserAvatarUrl,
-                        commitUserName: deployment.commitUserName,
-                        deploymentStatus: 'SCHEDULED',
-                      },
-                    },
-                  });
-
-                  await toast.promise(
-                    insertDeploymentPromise,
-                    {
-                      loading: 'Redeploying...',
-                      success: 'Redeployment has been started successfully.',
-                      error: 'An error occurred when redeploying your project.',
-                    },
-                    toastStyleProps,
-                  );
-                }}
-                startIcon={<ArrowCounterclockwiseIcon className="w-4 h-4" />}
-                className="rounded-full py-1 px-2 text-xs"
-              >
-                Redeploy
-              </Button>
-            </Tooltip>
-          )}
-
-          {isDeploymentLive && (
-            <div className="w-12 flex justify-end">
-              <Status status={StatusEnum.Live}>Live</Status>
-            </div>
-          )}
-
-          <div className="w-16 text-right font-mono text-sm- font-medium">
-            {deployment.commitSHA.substring(0, 7)}
-          </div>
-
-          {showTime && (
-            <div className="w-[80px] text-right font-mono text-sm- font-medium">
-              <AppDeploymentDuration
-                startedAt={deployment.deploymentStartedAt}
-                endedAt={deployment.deploymentEndedAt}
-              />
-            </div>
-          )}
-
-          <StatusCircle
-            status={deployment.deploymentStatus as DeploymentStatus}
-          />
-
-          <ChevronRightIcon className="h-4 w-4" />
-        </div>
-      </ListItem.Button>
-    </ListItem.Root>
-  );
-}
-
 interface OverviewDeploymentsProps {
   projectId: string;
   githubRepository: { fullName: string };
@@ -230,9 +69,18 @@ function OverviewDeployments({
     },
   });
 
-  if (loading) {
+  const {
+    data: scheduledOrPendingDeploymentsData,
+    loading: scheduledOrPendingDeploymentsLoading,
+  } = useScheduledOrPendingDeploymentsSubSubscription({
+    variables: {
+      appId: projectId,
+    },
+  });
+
+  if (loading || scheduledOrPendingDeploymentsLoading) {
     return (
-      <div style={{ height: '240px' }}>
+      <div className="h-60">
         <ActivityIndicator label="Loading deployments..." />
       </div>
     );
@@ -289,33 +137,29 @@ function OverviewDeployments({
     );
   }
 
-  const getLastLiveDeploymentId = getLastLiveDeployment(deployments);
-  const scheduledOrPendingDeploymentIndex = deployments.findIndex(
-    (deployment) =>
-      deployment.deploymentStatus === 'SCHEDULED' ||
-      deployment.deploymentStatus === 'PENDING',
-  );
+  const liveDeploymentId = getLastLiveDeployment(deployments);
+  const { deployments: scheduledOrPendingDeployments } =
+    scheduledOrPendingDeploymentsData;
 
   return (
-    <div className="rounded-x-lg flex flex-col divide-y-1 divide-gray-200 rounded-lg border border-veryLightGray">
-      {deployments.map((deployment, index) => {
-        const isDeploymentLive = deployment.id === getLastLiveDeploymentId;
-
-        return (
-          <OverviewDeployment
-            key={deployment.id}
-            deployment={deployment}
-            isDeploymentLive={isDeploymentLive}
-            showRedeploy={
-              scheduledOrPendingDeploymentIndex !== -1
-                ? scheduledOrPendingDeploymentIndex === index
-                : index === 0
-            }
-            disableRedeploy={scheduledOrPendingDeploymentIndex !== -1}
-          />
-        );
-      })}
-    </div>
+    <List className="rounded-x-lg flex flex-col divide-y-1 divide-gray-200 rounded-lg border border-veryLightGray">
+      {deployments.map((deployment, index) => (
+        <DeploymentListItem
+          key={deployment.id}
+          deployment={deployment}
+          isLive={deployment.id === liveDeploymentId}
+          showRedeploy={
+            scheduledOrPendingDeployments.length > 0
+              ? scheduledOrPendingDeployments.some(
+                  (scheduledOrPendingDeployment) =>
+                    scheduledOrPendingDeployment.id === deployment.id,
+                )
+              : index === 0
+          }
+          disableRedeploy={scheduledOrPendingDeployments.length > 0}
+        />
+      ))}
+    </List>
   );
 }
