+++ dashboard/src/gql/deployments/deployments.graphql
@@ -20,6 +20,17 @@ query getDeployments($id: uuid!, $limit: Int!, $offset: Int!) {
   }
 }
 
+subscription ScheduledOrPendingDeploymentsSub($appId: uuid!) {
+  deployments(
+    where: {
+      deploymentStatus: { _in: ["PENDING", "SCHEDULED"] }
+      appId: { _eq: $appId }
+    }
+  ) {
+    ...DeploymentRow
+  }
+}
+
 mutation InsertDeployment($object: deployments_insert_input!) {
   insertDeployment(object: $object) {
     ...DeploymentRow
